renaming folder to shared
