+++ packages/destination-actions/src/destinations/facebook-conversions-api/__tests__/shared-tests/app-data.test.ts
@@ -1,5 +1,5 @@
-import { generateAppData } from '../../refactor-files/functions'
-import { AnyPayload, GeneratedAppData } from '../../refactor-files/types'
+import { generateAppData } from '../../shared/functions'
+import { AnyPayload, GeneratedAppData } from '../../shared/types'
 
 describe('FacebookConversionsApi', () => {
   describe('AppData', () => {
