+++ packages/destination-actions/src/destinations/facebook-conversions-api/custom/index.ts
@@ -1,9 +1,9 @@
 import { ActionDefinition } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
-import { send } from '../refactor-files/functions'
-import { customFields } from '../refactor-files/fields'
-import { EventType } from '../refactor-files/constants'
+import { send } from '../shared/functions'
+import { customFields } from '../shared/fields'
+import { EventType } from '../shared/constants'
 
 const action: ActionDefinition<Settings, Payload> = {
   title: 'Custom Event',
