+++ packages/destination-actions/src/destinations/facebook-conversions-api/shared/functions.ts
